rename PortalComponent to Portal and move into atoms. Rename StickyContainer to Popper and move into atoms. Update tests & snapshot tests
