+++ src/index.tsx
@@ -12,6 +12,8 @@ export { default as Divider } from './atoms/Divider';
 export { default as Container } from './atoms/Container';
 export { default as Menu } from './atoms/Menu';
 export { default as MenuItem } from './atoms/MenuItem';
+export { default as Popper } from './atoms/Popper';
+export { default as Portal } from './atoms/Portal';
 export { default as Grid } from './atoms/Grid';
 export { default as Icon } from './atoms/Icon';
 export { default as Icons } from './atoms/Icon/registry';
